Folder restructured for profile and other pages to look better and readable.
