+++ Mobile Development CA2/Mobile Development CA2/Views/Components/DetailRowView.swift
@@ -1,0 +1,31 @@
+//
+//  DetailRowView.swift
+//  Mobile Development CA2
+//
+//  Created by Student on 28/03/2025.
+//
+
+import SwiftUI
+
+struct DetailRow: View {
+    let icon: String
+    let title: String
+    let value: String
+    
+    var body: some View {
+        HStack(alignment: .top, spacing: 15) {
+            Image(systemName: icon)
+                .foregroundColor(.blue)
+                .frame(width: 30)
+            
+            VStack(alignment: .leading, spacing: 5) {
+                Text(title)
+                    .font(.headline)
+                Text(value)
+                    .font(.body)
+                    .foregroundColor(.secondary)
+            }
+        }
+        .padding(.vertical, 8)
+    }
+}
